+++ CHANGELOG.md
@@ -1,8 +1,13 @@
 Releases:
 
+V7
+==============
+- Dock more than one ship on a planet (with a threshold 75%)
+
 V6
 ==============
 - Do not thrust toward planet if another ship is already headed to the same planet
+- Weakness: Only docking one ship per planet
 
 V5
 ==============